+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/environment-check
 
+## 0.15.45
+
+### Patch Changes
+
+-   24e45780: Updated dependency: axios@1.4.0
+-   Updated dependencies [24e45780]
+    -   @sap-ux/axios-extension@1.4.1
+    -   @sap-ux/btp-utils@0.11.8
+
 ## 0.15.44
 
 ### Patch Changes